Move the web server to its own module
